Новая папка (2)/pr/pyproject.toml
[project]
name = "pr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
